% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diviana-package.R
\docType{package}
\name{diviana-package}
\alias{diviana}
\alias{diviana-package}
\title{diviana: Shiny App for Disaster Victim Identification}
\description{
A 'shiny' application for disaster victim identification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/diviana}
  \item Report bugs at \url{https://github.com/magnusdv/diviana/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
