% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch DIVIANA}
\usage{
launchApp()
}
\value{
No return value, called for side effects.
}
\description{
This launches the DIVIANA app.
}
\examples{

\dontrun{
launchApp()
}

}
